+++ src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/TestApFalsePattern.kt
@@ -18,13 +18,13 @@ class TestApFalsePattern(
 
     init {
         cg_map.entries.forEach { (key, value) ->
-            if (cg_regex0.matches(key)) {
+            if (cg_regex0.containsMatchIn(key)) {
                 require(value is Int) { "$key is not the correct type, expecting Int" }
                 require(value in 0..99) { "$key not in range 0..99 - $value" }
             }
         }
         cg_map.keys.forEach { key ->
-            if (!cg_regex0.matches(key))
+            if (!cg_regex0.containsMatchIn(key))
                 throw IllegalArgumentException("Unexpected field $key")
         }
     }